requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2
scipy==1.11.4
flasgger==0.9.7.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
python-json-logger==2.0.7
